+++ .eslintrc.js
@@ -87,5 +87,6 @@ module.exports = {
     "consistent-return": "off",
     "no-unused-vars": "off",
     "@typescript-eslint/no-unused-vars": ["error", { argsIgnorePattern: "^_" }],
+    "react/prop-types": "off",
   },
 };
