+++ src/ui/BottomDrawer/index.ts
@@ -1,3 +1,3 @@
-export * from "./SelectBottomDrawer";
 export * from "./ProfileBottomDrawer";
 export { default as BottomDrawer } from "./BottomDrawer";
+export { default as PullToCloseBottomDrawer } from "./PullToCloseBottomDrawer";
