+++ src/ui/Select/Select.stories.tsx
@@ -1,26 +1,35 @@
-import { ComponentStory, ComponentMeta } from "@storybook/react";
+import { useState } from "react";
 
-import { colors } from "../tokens/color";
+import { ComponentStory, ComponentMeta } from "@storybook/react";
 
 import Select from "./Select";
 
 export default {
-  argTypes: {
-    color: { control: "select", options: Object.keys(colors) },
-  },
   component: Select,
   title: "UI/Select",
 } as ComponentMeta<typeof Select>;
 
-const Template: ComponentStory<typeof Select> = (args) => <Select {...args} />;
+const Template: ComponentStory<typeof Select> = ({ size, title }) => {
+  const [value, setValue] = useState("");
+  return (
+    <Select
+      options={[
+        { label: "test1", value: "testvalue1" },
+        { label: "test2", value: "testvalue2" },
+      ]}
+      placeholder="trigger"
+      size={size}
+      title={title}
+      value={value}
+      onValueChange={(v) => setValue(v)}
+    />
+  );
+};
 
 export const Default = Template.bind({});
-Default.args = {
-  placeholder: "Select",
-};
+Default.args = {};
 
 export const Small = Template.bind({});
 Small.args = {
-  placeholder: "Select",
   size: "small",
 };
